requirements.txt
PyPDF2==3.0.1
openai>=1.6.0
python-dotenv==1.0.0
httpx>=0.25.0 